Refactor: Boast컴포넌트 분리
