+++ src/components/PostWrite/PostWrite.jsx
@@ -2,9 +2,13 @@ import React, { useState } from 'react';
 import { styled } from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import Select from './Select';
+import { useSelector } from 'react-redux';
 
 function PostWrite() {
   const [fileName, setFileName] = useState('');
+  const user = useSelector(state => {
+    return state.user;
+  });
   const navigate = useNavigate();
   const cancelWrite = () => {
     navigate(-1);
@@ -13,13 +17,19 @@ function PostWrite() {
     const name = target.value.split('/').pop().split('\\').pop();
     setFileName(name);
   };
+  const onSubmitHandler = (e) => {
+    e.preventDefault();
+    // console.log(e)
+    const {title, content, img} = e.target.name;
+    // console.log(title, content, img)
+  }
   return (
-    <StForm>
+    <StForm onSubmit={onSubmitHandler}>
       <FormHeader>
         <Select />
-        <StInput placeholder="제목을 입력하세요"></StInput>
+        <StInput placeholder="제목을 입력하세요" name='title'></StInput>
       </FormHeader>
-      <StyledTextarea placeholder="내용을 입력하세요" rows="30" cols="118"></StyledTextarea>
+      <StyledTextarea placeholder="내용을 입력하세요" name='content' rows="30" cols="118"></StyledTextarea>
       <FormBottom>
         <FileField>
           <FileLabel>
@@ -27,6 +37,7 @@ function PostWrite() {
             <input
               type="file"
               accept="image/gif, image/jpeg, image/png"
+              name='img'
               style={{ display: 'none' }}
               onChange={fileHandler}
             />
@@ -34,7 +45,7 @@ function PostWrite() {
           <File>{fileName}</File>
         </FileField>
         <ButtonArea>
-          <Button>작성하기</Button>
+          <Button type='submit'>작성하기</Button>
           <Button onClick={cancelWrite} type="button">
             취소하기
           </Button>
