+++ src/shared/Router.jsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Login from '../pages/Login';
+import LoginPage from '../pages/Login';
 import Home from '../pages/Home';
-import Boast from '../pages/Boast';
+import BoastPage from '../pages/Boast';
 import Qna from '../pages/Qna';
 import Join from '../pages/Join';
 import PostWrite from '../pages/PostWrite';
@@ -13,8 +13,8 @@ const Router = () => {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/join" element={<Join />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/boast" element={<Boast />} />
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/boast" element={<BoastPage />} />
         <Route path="/qna" element={<Qna />} />
         <Route path="/postWrite" element={<PostWrite />} />
         <Route path="/detailPage" element={<DetailPage />} />
